Simplify by making MOGReducers blocks
In order to support processes were values are accumulated in the process
to be passed on in chunks or when the process finish MOGReducer was made
into an object with initial, reduce and complete blocks.

When the process is finished complete was called to allow the process to
flush any internal buffers. While this makes it more powerful it also
makes makes the implementation more complex and adds to the concepts
required to be known by the user.

Another drawback is that it makes it more cumbersome to supply a user
defined reducer since an object needs to be created with the reduce
block set.

While the last paragraph is a benefit of dropping the MOGReducer class
and simply make it into a `(C, A) -> A` block, using a class had the
benefit of make the accumulator type opaque for the user of the shipped
reducers (MOGArrayReducer etc).